refac: moved pages to src
