+++ tests/home.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import Home from '@pages/index'
+import Home from 'src/pages/index'
 
 describe('Home Page', () => {
   it('should render welcome message', () => {
