+++ tsconfig.json
@@ -8,7 +8,7 @@
       "@styles": ["styles/global.css"],
       "@prisma": ["prisma/prisma.ts"],
       "@lib": ["lib/index.ts"],
-      "@pages/*": ["pages/*"]
+      "@pages/*": ["src/pages/*"]
     },
     "target": "es5",
     "lib": ["dom", "dom.iterable", "esnext"],
